backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
alembic
pytest
httpx
